+++ bin/predicate/src/cli/mod.rs
@@ -3,7 +3,7 @@ use clap::Parser;
 use clap::Subcommand;
 
 use crate::commands::common::CommonCommands;
-use crate::commands::{eq, lt, lte};
+use crate::commands::{eq, gt, gte, lt, lte};
 
 #[derive(Debug, Parser)]
 #[command(author, version, about = "AFS Predicate CLI")]
@@ -18,13 +18,21 @@ pub enum CliCommand {
     #[command(name = "eq", about = "Get data in a Table that is equal to the input")]
     /// Get data in a Table that is equal to the input
     Eq(eq::EqCommand),
-    // #[command(name = "gt", about = "Get data in a Table that is greater than the input")]
-    // /// Get data in a Table that is greater than the input
-    // Gt(gt::GtCommand),
 
-    // #[command(name = "gte", about = "Get data in a Table that is greater than or equal to the input")]
-    // /// Get data in a Table that is greater than or equal to the input
-    // Gte(gte::GteCommand),
+    #[command(
+        name = "gt",
+        about = "Get data in a Table that is greater than the input"
+    )]
+    /// Get data in a Table that is greater than the input
+    Gt(gt::GtCommand),
+
+    #[command(
+        name = "gte",
+        about = "Get data in a Table that is greater than or equal to the input"
+    )]
+    /// Get data in a Table that is greater than or equal to the input
+    Gte(gte::GteCommand),
+
     #[command(name = "lt", about = "Get data in a Table that is less than the input")]
     /// Get data in a Table that is less than the input
     Lt(lt::LtCommand),
@@ -49,6 +57,30 @@ impl Cli {
                         value: eq.args.value.clone(),
                         output_file: eq.args.output_file.clone(),
                         silent: eq.args.silent,
+                    },
+                };
+                cmd.execute(config).unwrap();
+            }
+            CliCommand::Gt(gt) => {
+                let cmd = gt::GtCommand {
+                    args: CommonCommands {
+                        db_file_path: gt.args.db_file_path.clone(),
+                        table_id: gt.args.table_id.clone(),
+                        value: gt.args.value.clone(),
+                        output_file: gt.args.output_file.clone(),
+                        silent: gt.args.silent,
+                    },
+                };
+                cmd.execute(config).unwrap();
+            }
+            CliCommand::Gte(gte) => {
+                let cmd = gte::GteCommand {
+                    args: CommonCommands {
+                        db_file_path: gte.args.db_file_path.clone(),
+                        table_id: gte.args.table_id.clone(),
+                        value: gte.args.value.clone(),
+                        output_file: gte.args.output_file.clone(),
+                        silent: gte.args.silent,
                     },
                 };
                 cmd.execute(config).unwrap();
