+++ bin/predicate/src/commands/mod.rs
@@ -1,4 +1,6 @@
 pub mod common;
 pub mod eq;
+pub mod gt;
+pub mod gte;
 pub mod lt;
 pub mod lte;
